processor/requirements.txt
Werkzeug==2.2.2
Flask==2.2.3
cloudevents==1.11.0
phidata==2.5.32
ollama==0.4.0